+++ src/constants/board.js
@@ -1,0 +1,7 @@
+export const STORAGE_KEY = 'dynamic-kanban.tasks'
+
+export const COLUMNS = [
+  { key: 'todo', label: 'To Do' },
+  { key: 'in-progress', label: 'In Progress' },
+  { key: 'done', label: 'Done' },
+]
